requirements.txt
Django~=3.1.3
django-crispy-forms==1.10.0
numpy==1.19.4
psycopg2==2.8.6
jsonpickle==1.5.0
scipy==1.6.0
elasticsearch==7.10.1
networkx~=2.5
matplotlib~=3.3.3
requests~=2.25.1